refactor: move ProfileImage to common dir
